requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiocache==0.12.2
    # via -r requirements.in
aiosqlite==0.17.0
    # via tortoise-orm
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[all]==0.111.1
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
iso8601==1.1.0
    # via tortoise-orm
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.6
    # via fastapi
pydantic==2.8.2
    # via
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
    #   tortoise-orm
pydantic-core==2.20.1
    # via pydantic
pydantic-extra-types==2.9.0
    # via fastapi
pydantic-settings==2.4.0
    # via fastapi
pygments==2.18.0
    # via rich
pypika-tortoise==0.1.6
    # via tortoise-orm
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via tortoise-orm
pyyaml==6.0.1
    # via
    #   fastapi
    #   uvicorn
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
tortoise-orm==0.21.5
    # via -r requirements.in
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typer
ujson==5.10.0
    # via fastapi
uvicorn[standard]==0.30.3
    # via fastapi
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
